requirements.txt
# Core Dependencies
fastapi
uvicorn
pydantic
pydantic-settings
python-dotenv

# MCP SDK
mcp

# Lead Generation & Data
faker
sqlalchemy

# AI/LLM
groq

# Email
aiosmtp

# Utilities
python-multipart
aiofiles
httpx

# CORS
fastapi-cors